+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/tasks/DelayedAuthHook.java
@@ -8,8 +8,9 @@ import com.github.games647.fastlogin.bukkit.hooks.LoginSecurityHook;
 import com.github.games647.fastlogin.bukkit.hooks.UltraAuthHook;
 import com.github.games647.fastlogin.bukkit.hooks.xAuthHook;
 import com.github.games647.fastlogin.core.hooks.AuthPlugin;
-import com.google.common.collect.Lists;
 
+import java.lang.reflect.Constructor;
+import java.util.Arrays;
 import java.util.List;
 
 import org.bukkit.Bukkit;
@@ -41,28 +42,7 @@ public class DelayedAuthHook implements Runnable {
     }
 
     private boolean registerHooks() {
-        AuthPlugin<Player> authPluginHook = null;
-
-        try {
-            @SuppressWarnings("unchecked")
-            List<Class<? extends AuthPlugin<Player>>> supportedHooks = Lists.newArrayList(AuthMeHook.class
-                    , CrazyLoginHook.class, LogItHook.class, LoginSecurityHook.class, UltraAuthHook.class
-                    , xAuthHook.class);
-
-            for (Class<? extends AuthPlugin<Player>> clazz : supportedHooks) {
-                String pluginName = clazz.getSimpleName().replace("Hook", "");
-                //uses only member classes which uses AuthPlugin interface (skip interfaces)
-                if (Bukkit.getPluginManager().isPluginEnabled(pluginName)) {
-                    //check only for enabled plugins. A single plugin could be disabled by plugin managers
-                    authPluginHook = clazz.newInstance();
-                    plugin.getLog().info("Hooking into auth plugin: {}", pluginName);
-                    break;
-                }
-            }
-        } catch (InstantiationException | IllegalAccessException ex) {
-            plugin.getLog().error("Couldn't load the integration class", ex);
-        }
-
+        AuthPlugin<Player> authPluginHook = getAuthHook();
         if (authPluginHook == null) {
             //run this check for exceptions (errors) and not found plugins
             plugin.getLog().warn("No support offline Auth plugin found. ");
@@ -70,9 +50,42 @@ public class DelayedAuthHook implements Runnable {
         }
 
         if (plugin.getCore().getAuthPluginHook() == null) {
+            plugin.getLog().info("Hooking into auth plugin: {}", authPluginHook.getClass().getSimpleName());
             plugin.getCore().setAuthPluginHook(authPluginHook);
         }
 
         return true;
+    }
+
+    private AuthPlugin<Player> getAuthHook() {
+        try {
+            @SuppressWarnings("unchecked")
+            List<Class<? extends AuthPlugin<Player>>> hooks = Arrays.asList(AuthMeHook.class,
+                    CrazyLoginHook.class, LogItHook.class, LoginSecurityHook.class, UltraAuthHook.class,
+                    xAuthHook.class);
+
+            for (Class<? extends AuthPlugin<Player>> clazz : hooks) {
+                String pluginName = clazz.getSimpleName().replace("Hook", "");
+                //uses only member classes which uses AuthPlugin interface (skip interfaces)
+                if (Bukkit.getPluginManager().isPluginEnabled(pluginName)) {
+                    //check only for enabled plugins. A single plugin could be disabled by plugin managers
+                    return newInstance(clazz);
+                }
+            }
+        } catch (ReflectiveOperationException ex) {
+            plugin.getLog().error("Couldn't load the integration class", ex);
+        }
+
+        return null;
+    }
+
+    private AuthPlugin<Player> newInstance(Class<? extends AuthPlugin<Player>> clazz)
+            throws ReflectiveOperationException {
+        try {
+            Constructor<? extends AuthPlugin<Player>> cons = clazz.getConstructor(FastLoginBukkit.class);
+            return cons.newInstance(plugin);
+        } catch (NoSuchMethodException noMethodEx) {
+            return clazz.newInstance();
+        }
     }
 }