+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### 1.11
 
+* Use direct proxies instead of ssl factories for multiple IP-addresses
+* Remove local address check for multiple IP-addresses
+* Fix parsing of local IP-addresses
+* Fix address rotating for contacting the Mojang API
 * Optimize issue template
 * Use Instant for timestamps
 * Migrate SLF4J logging (Fixes #177)